modified functions and header. Added Makefile.

+++ libft/ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-size_t	ft_strlen(char *str)
+int	ft_strlen(char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i] != '\0')
